chore: clean up catalogers
