+++ syft/pkg/cataloger/file/test-fixtures/version.txt
